+++ DataPersistence/src/main/java/sep3/project/daos/impl/PostDatabase.java
@@ -49,7 +49,7 @@ public class PostDatabase implements PostPersistence {
 	}
 
 	@Override
-	public void deletePost(int id) throws SQLException {
+	public void deletePost(long id) throws SQLException {
 		Connection connection = DBConnection.getConnection();
 
 		try
@@ -60,14 +60,14 @@ public class PostDatabase implements PostPersistence {
 							"(SELECT id FROM comment WHERE post_id IN " +
 							"(SELECT id FROM post WHERE id = ?))"
 			);
-			statement_sub_comments.setInt(1, id);
+			statement_sub_comments.setLong(1, id);
 			statement_sub_comments.execute();
 			//deleting comments
 			PreparedStatement statement_comments = connection.prepareStatement(
 					"DELETE FROM comment WHERE post_id IN " +
 							"(SELECT id FROM post WHERE id = ?)"
 			);
-			statement_comments.setInt(1, id);
+			statement_comments.setLong(1, id);
 			statement_comments.execute();
 
 			//deleting likes
@@ -75,7 +75,7 @@ public class PostDatabase implements PostPersistence {
 					"DELETE FROM likepost WHERE post_id IN " +
 							"(SELECT id FROM post WHERE id = ?)"
 			);
-			statement_likes.setInt(1, id);
+			statement_likes.setLong(1, id);
 			statement_likes.execute();
 
 			//deleting reports
@@ -83,7 +83,7 @@ public class PostDatabase implements PostPersistence {
 					"DELETE FROM report WHERE post_id IN " +
 							"(SELECT id FROM post WHERE id = ?)"
 			);
-			statement_reports.setInt(1, id);
+			statement_reports.setLong(1, id);
 			statement_reports.execute();
 
 			//deleting from tags
@@ -94,14 +94,14 @@ public class PostDatabase implements PostPersistence {
 							"(SELECT id FROM post WHERE id = ?)"
 			);
 
-			statement_tags.setInt(1, id);
+			statement_tags.setLong(1, id);
 			statement_tags.execute();
 			// deleting ban_post
 			PreparedStatement statement_ban = connection.prepareStatement(
 					"DELETE FROM ban_post WHERE post_id IN " +
 							"(SELECT id FROM post WHERE id = ?);"
 			);
-			statement_ban.setInt(1, id);
+			statement_ban.setLong(1, id);
 			statement_ban.execute();
 
 			//deleting post
@@ -109,7 +109,7 @@ public class PostDatabase implements PostPersistence {
 					"DELETE FROM post " +
 							"WHERE id = ?"
 			);
-			statement.setInt(1, id);
+			statement.setLong(1, id);
 			statement.execute();
 		}
 		finally {
@@ -118,7 +118,7 @@ public class PostDatabase implements PostPersistence {
 	}
 
 	@Override
-	public ArrayList<PostData> getPost(int id, long userId, String titleContains) throws SQLException {
+	public ArrayList<PostData> getPost(long id, long userId, String titleContains) throws SQLException {
 		Connection connection = DBConnection.getConnection();
 		ArrayList<PostData> posts = new ArrayList<>();
 		ResponseGetUsers response = null;
@@ -219,14 +219,14 @@ public class PostDatabase implements PostPersistence {
 		}
 	}
 
-	private ResultSet getById(Connection connection, int id) {
+	private ResultSet getById(Connection connection, long id) {
 		PreparedStatement statement = null;
 
 		try {
 			statement = connection.prepareStatement(
 					"SELECT * FROM \"post\" WHERE id = ?");
 
-			statement.setInt(1, id);
+			statement.setLong(1, id);
 			return statement.executeQuery();
 
 		} catch (SQLException e) {
@@ -248,14 +248,14 @@ public class PostDatabase implements PostPersistence {
 			throw new RuntimeException(e);
 		}
 	}
-	private ResultSet getByByUserIdAndPostId(Connection connection, long userId, int id) {
+	private ResultSet getByByUserIdAndPostId(Connection connection, long userId, long id) {
 		PreparedStatement statement = null;
 
 		try {
 			statement = connection.prepareStatement(
 					"SELECT * FROM \"post\" WHERE user_id = ? AND id=?");
 			statement.setLong(1, userId);
-			statement.setInt(2, id);
+			statement.setLong(2, id);
 			return statement.executeQuery();
 
 		} catch (SQLException e) {
@@ -278,13 +278,13 @@ public class PostDatabase implements PostPersistence {
 		}
 	}
 
-	private ResultSet getByIdAndTitle(Connection connection, int id, String title) {
+	private ResultSet getByIdAndTitle(Connection connection, long id, String title) {
 		PreparedStatement statement = null;
 
 		try {
 			statement = connection.prepareStatement(
 					"SELECT * FROM \"post\" WHERE id = ? AND lower(title) LIKE '%' || ? || '%'");
-			statement.setInt(1, id);
+			statement.setLong(1, id);
 			statement.setString(2, title);
 			return statement.executeQuery();
 
@@ -292,14 +292,14 @@ public class PostDatabase implements PostPersistence {
 			throw new RuntimeException(e);
 		}
 	}
-	private ResultSet getByAllParameters(Connection connection, int id, long userId, String title)
+	private ResultSet getByAllParameters(Connection connection, long id, long userId, String title)
 	{
 		PreparedStatement statement = null;
 
 		try {
 			statement = connection.prepareStatement(
 					"SELECT * FROM \"post\" WHERE id = ? AND lower(title) LIKE '%' || ? || '%' AND user_id = ?");
-			statement.setInt(1, id);
+			statement.setLong(1, id);
 			statement.setString(2, title);
 			statement.setLong(3, userId);
 			return statement.executeQuery();
@@ -311,7 +311,7 @@ public class PostDatabase implements PostPersistence {
 
 	private PostData createPostFromQuery(ResultSet resultSet) throws SQLException {
 		return PostData.newBuilder()
-				.setId(resultSet.getInt("id"))
+				.setId(resultSet.getLong("id"))
 				.setUserId(resultSet.getLong("user_id"))
 				.setDescription(resultSet.getString("description"))
 				.setTitle(resultSet.getString("title"))
@@ -322,7 +322,7 @@ public class PostDatabase implements PostPersistence {
 	}
 
 	@Override
-	public ResponseUpdatePost updatePost(int id, String title, String description, String[] tags) throws SQLException {
+	public ResponseUpdatePost updatePost(long id, String title, String description, String[] tags) throws SQLException {
 		Connection connection = DBConnection.getConnection();
 		ResponseUpdatePost response = null;
 		try
@@ -330,13 +330,13 @@ public class PostDatabase implements PostPersistence {
 			PreparedStatement statement_text = connection.prepareStatement("UPDATE post SET title = ?, description = ? WHERE id = ?;");
 			statement_text.setString(1, title);
 			statement_text.setString(2, description);
-			statement_text.setInt(3, id);
+			statement_text.setLong(3, id);
 
 			statement_text.execute();
 
 			//first deleting
 			PreparedStatement statement_tags_delete = connection.prepareStatement("DELETE FROM post_tag WHERE post_id IN (SELECT id FROM post WHERE id = ?);");
-			statement_tags_delete.setInt(1, id);
+			statement_tags_delete.setLong(1, id);
 			statement_tags_delete.execute();
 
 			//then inserting
@@ -347,7 +347,7 @@ public class PostDatabase implements PostPersistence {
 				statement1.execute();
 
 				PreparedStatement statement = connection.prepareStatement("INSERT INTO post_tag(post_id, tag_name) VALUES (?, ?)");
-				statement.setInt(1, id);
+				statement.setLong(1, id);
 				statement.setString(2, tag);
 				statement.execute();
